10番終了

+++ src/main/java/com/example/repository/ItemRepository.java
@@ -3,7 +3,9 @@ import  org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ItemList;
 
-public interface ItemRepository extends JpaRepository<ItemList, Long>{
+public interface ItemRepository extends JpaRepository<ItemList, Integer>{
+
+	
 	
 	
 }
